Restructured project layout.
There were getting to be a few to many files in one directory for my
liking, so I restructured the file layout of the project.

Tests went into 'tests', cli stuff into 'cli'.
